refactor: Moved documents.py from model to src
